requirements.txt
numpy
torch
torchvision
Pillow
tifffile
scipy
matplotlib
tensorboardX
